SDK: rename block entryfiles and add documentation
