+++ client/gutenberg/extensions/editor-notes/editor.scss
