+++ client/gutenberg/extensions/markdown/editor.jsx
@@ -11,7 +11,7 @@ import { registerBlockType } from '@wordpress/blocks';
 /**
  * Internal dependencies
  */
-import './style.scss';
+import './editor.scss';
 import edit from './edit';
 import save from './save';
