+++ client/gutenberg/extensions/presets/jetpack/editor.js
@@ -5,4 +5,4 @@
  */
 import './utils/block-category'; // Register the Jetpack category
 import 'gutenberg/extensions/markdown/editor';
-import 'gutenberg/extensions/tiled-gallery/editor.js';
+import 'gutenberg/extensions/tiled-gallery/editor';
