+++ client/gutenberg/extensions/presets/o2/view.js
@@ -3,4 +3,4 @@
 /**
  * Internal dependencies
  */
-import 'gutenberg/extensions/trial/view.js';
+import 'gutenberg/extensions/trial/view';
