+++ client/gutenberg/extensions/prev-next/editor.scss
