+++ client/gutenberg/extensions/publicize/no-connections.jsx
@@ -79,4 +79,3 @@ class PublicizeNoConnections extends Component {
 }
 
 export default PublicizeNoConnections;
-
