+++ client/gutenberg/extensions/tiled-gallery/editor.js
@@ -9,7 +9,7 @@ import { createBlock, registerBlockType } from '@wordpress/blocks';
 /**
  * Internal dependencies
  */
-import './style.scss';
+import './view.scss';
 import TiledGalleryEdit from './edit.jsx';
 import TiledGallerySave from './save.jsx';
