+++ client/gutenberg/extensions/tiled-gallery/view.js
@@ -3,4 +3,4 @@
 /**
  * Internal dependencies
  */
-import './style.scss';
+import './view.scss';
