+++ client/gutenberg/extensions/tiled-gallery/view.scss
@@ -1,20 +1,20 @@
 .wp-block-a8c-tiled-gallery {
-	.tiled-gallery__square {
+	.tiled-gallery-square {
 		clear: both;
 		margin: 0 0 20px;
 		overflow: hidden;
 	}
-	.tiled-gallery__image {
+	.tiled-gallery-image {
 		margin: 2px !important; /* Ensure that this value isn't overridden by themes that give content images blanket margins */
 	}
-	.tiled-gallery__group {
+	.gallery-group {
 		float: left;
 		position: relative;
 	}
-	.tiled-gallery__row {
+	.gallery-row {
 		 overflow: hidden;
 	}
-	.tiled-gallery__item {
+	.tiled-gallery-item {
 		float: left;
 		margin: 0;
 		position: relative;
@@ -31,8 +31,8 @@
 		figure {
 			margin: 0;
 		}
-		.tiled-gallery__image,
-		.tiled-gallery__image:hover { /* Needs to reset some properties for theme compatibility */
+		.tiled-gallery-image,
+		.tiled-gallery-image:hover { /* Needs to reset some properties for theme compatibility */
 			background: none;
 			border: none;
 			box-shadow: none;
@@ -42,7 +42,7 @@
 		}
 	}
 
-	.tiled-gallery__caption { /* Captions */
+	.tiled-gallery-caption { /* Captions */
 		background: #eee;
 		background: rgba( 255,255,255,0.8 );
 		color: #333;
@@ -58,7 +58,7 @@
 		white-space: nowrap;
 	}
 
-	.tiled-gallery__item-small .tiled-gallery__caption { /* Smaller captions */
+	.tiled-gallery-item-small .tiled-gallery-caption { /* Smaller captions */
 		font-size: 11px;
 	}
 }
